requirements.txt
Django==5.2
wagtail==6.4
psycopg2-binary==2.9.10
django-storages[s3]==1.14.4
boto3==1.35.0
gunicorn==23.0.0
dj-database-url==2.3.0
python-decouple==3.8
Pillow==11.0.0
whitenoise==6.8.2
GDAL==3.10.0
